move apply to utils and remove entirely the compatibility module
