+++ .coveragerc
@@ -1,6 +1,5 @@
 [run]
 omit =
     */test_*.py
-    datashader/compatibility.py
 source =
     datashader
