+++ datashader/tests/test_utils.py
@@ -1,7 +1,7 @@
 from __future__ import absolute_import
 from datashape import dshape
 
-from datashader.utils import Dispatcher, isreal
+from datashader.utils import Dispatcher, apply, isreal
 
 
 def test_Dispatcher():
@@ -26,3 +26,8 @@ def test_isreal():
     assert isreal('float64')
     assert not isreal('complex64')
     assert not isreal('{x: int64, y: float64}')
+
+def test_apply():
+    f = lambda a, b, c=1, d=2: a + b + c + d
+    assert apply(f, (1, 2,)) == 6
+    assert apply(f, (1, 2,), dict(c=3)) == 8
